+++ include/cassette/cgui-screen.h
@@ -34,10 +34,10 @@ extern "C" {
 
 struct cgui_screen
 {
-	 int16_t x;
-	 int16_t y;
-	uint16_t width;
-	uint16_t height;
+	double x;
+	double y;
+	double width;
+	double height;
 	bool primary;
 };
