+++ include/cassette/cgui-zone.h
@@ -36,10 +36,10 @@ extern "C" {
 struct cgui_zone
 {
 	cairo_t *drawable;
-	 int16_t x;
-	 int16_t y;
-	uint16_t width;
-	uint16_t height;
+	double x;
+	double y;
+	double width;
+	double height;
 };
 
 /************************************************************************************************************/
